Move title() from utilities.utils to utilities.string
